Add temporary indexer to trigger refetching of blocks where internal transactions number doesn't match
Problem: There are inconsistencies in the database regarding blocks and their internal transactions' `block_number`.
A fix for new data is already mergeed and the first and last block affected are known, but they have to be corrected.

Solution: implement a temporary indexer that handles this range of blocks, triggering their refetch.