Add temporary indexer to trigger refetching of blocks with missing token transfers
Problem: There are inconsistencies in the database regarding blocks with missing token transfers.
A fix for new data is already merged and the blocks affected are known, but they have to be corrected.

Solution: implement a temporary indexer that handles this blocks (whose number are stored into a temporary table), triggering thir refetch.